requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.0            # via django
beautifulsoup4==4.9.3     # via django-bootstrap4
django-bootstrap4==2.3.1  # via -r requirements.in
django-filter==2.4.0      # via -r requirements.in
django==3.0.3             # via -r requirements.in, django-bootstrap4, django-filter, djangorestframework
djangorestframework==3.12.1  # via -r requirements.in
importlib-metadata==2.0.0  # via django-bootstrap4
pytz==2020.1              # via django
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.4.1           # via django
zipp==3.4.0               # via importlib-metadata
